Word v2.5.0
This is an update to the Word schema to add `startTime` and `endTime` attributes to both the Word and Phone schemas. These properties are useful for phonetic data analysis. Unlike the `startTime` and `endTime` properties on the Sentence, these properties can occur individually. A `startTime` can be included without also including an `endTime`. This is useful when the endTime of one word token / phone is also the startTime of another word token / phone, so specifying both would be redundant.

## Changelog
* Word v2.5.0 (closes #198)
* NEW: Word.startTime
* NEW: Word.endTime
* NEW: Word.phonemes.startTime
* NEW: Word.phonemes.endTime