use official GeoJSON schema (#228)
This update removes the DLx version of the GeoJSON schema, and uses the [official version of the schema](https://github.com/geojson/schema) instead. It is not currently possible to point to the URL of the official GeoJSON schema directly, because it adheres to JSON Schema Draft 07.

This PR does **not** close issue #228.