Utterance 5.0.0
* CHANGE: Remove translationType field (closes #263)

Removed translationType from translation schema, Changed description in the translation schema and in the Utterance.translation description to reflect change (clarifying that translation should be used for free translations only). Updated translation example to reflect changes. Fixed a couple typos in Utterance description.

* NEW: Add Utterance.literalTranslation (closes #262)

Added literalTranslation field to Utterance. The field takes a Translation object as the optional literal translation. A free translation is already required for an Utterance.

* NEW: Add Utterance.phonetic (closes #272)

Created Utterance.phonetic field for the phonetic transcription of an Utterance. Currently it takes any string, however in the future this should be restricted to IPA by a JSON pattern. Patched title field of literalTranslation.

* VERSION: Utterance 5.0.0

* VERSION: Translation 3.0.0

Co-Authored-By: vadekh <36901345+vadekh@users.noreply.github.com>